Pass slice to a partition() function
Since slice in golang is reference type, when passing slice to a
function, pointer to slice header is passed to function and it is
'passed by value'. Slice header contains length and capacity information
and pointer to array that holds slice values. It's not necessary to pass
the address of the slice pointer.

This change introduced another -19% in terms of speed of
quartilePoints() function:

benchmark                     old ns/op     new ns/op     delta
BenchmarkPearson-4            1.33          1.37          +3.01%
BenchmarkFillBuckets-4        53701         51675         -3.77%
BenchmarkQuartilePoints-4     15681         12569         -19.85%
BenchmarkHash-4               78243         72349         -7.53%
BenchmarkModDiff-4            1.35          1.37          +1.48%
BenchmarkDigestDistance-4     195           174           -10.77%
BenchmarkDiffTotal-4          232           230           -0.86%